modules/api-locations-producer/requirements.txt
# Flask-based service dependencies
aniso8601==7.0.0
attrs==19.1.0
click>=8.0
Flask==2.1.2
flask_accepts
flask-cors==3.0.8
Flask-RESTful==0.3.7
flask-restx
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
GeoAlchemy2==0.8.4
itsdangerous>=2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
jsonschema==3.0.2
werkzeug==2.2.3

# gRPC and Kafka dependencies
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.3
kafka-python==2.0.6

# Other dependencies
marshmallow==3.17.0
marshmallow-sqlalchemy==0.23.1
#psycopg2==2.9.5  # or psycopg2-binary==2.9.5
psycopg2-binary==2.9.5
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
shapely==1.7.0
SQLAlchemy==1.4.41
requests==2.31.0
six==1.16.0
